(releng) Set PORTEPOCH for ports with a version going backward.
We now use the file date instead of the current date for building the ports version to ease up maintenance.  So some ports had to be updated accordingly.

Note for myself: do not plan to use the output of 'svn diff' on the local repository after running 'svn commit'.  meh.